requirements.txt
fastapi
uvicorn[standard]
pydantic
pymongo
bcrypt
python-jose[cryptography]
python-dotenv